add visualizations
